transaction page as function and starfx store
